face-verification-api/requirements.txt
flask==3.1.0
flask-cors==5.0.0
mtcnn==1.0.0
keras-facenet==0.3.2
opencv-python==4.10.0.84
scipy==1.14.1
Pillow==11.0.0
numpy==1.26.4
torch==2.1.0
torchvision==0.16.0
git+https://github.com/openai/CLIP.git
pillow==10.1.0
tensorflow==2.15.0
